Refactored internal logic of InclusionService into InclusionResolvers
This made it possible to further improve performance by only activating
the InclusionResolvers that have been actually configured to include or
exclude nodes. It also made it possible for custom inclusion resolvers
to be every bit as powerful as the built-in ones, because they are all
treated the same.

The InclusionResolver mechanism also allowed me to get rid of the
InclusionNode and replace it with ValueNode. The original logic that
justified the creation of the InclusionNode has been migrated to the
NodePathInclusionResolver.