Reimplemented property access exception handling
I removed the write exception handler method from the exception handler
interface because it could only be triggered when DiffNode#canonicalSet
or DiffNode#set get explicitly invoked by the user of the library.

But at that point the exception could simply be caught and handled
directly, so the indirection via exception handler seems unneccessary.

[Fixes #116]