CollectionItemElementSelector now returns a constant hashCode
The ElementSelector `hashCode` is only needed to use it as key in a
Map. With introduction of the `IdentityStrategy` this adds an
unnecessary source of confusion and complexity when implementing custom
IdentityStrategies.

To avoid this, returning a constant hashCode seems like a small price
to pay. Yes, it may have a small performance impact, but we can still
optimize when that turns out to be a problem.